+++ packages/super-editor/src/core/super-converter/exporter.js
@@ -1731,17 +1731,25 @@ function prepareCheckboxAnnotation(params) {
  */
 function prepareHtmlAnnotation(params) {
   const {
-    node: { attrs = {} },
+    node: { attrs = {}, marks = [] },
+    editorSchema,
   } = params;
 
   const parser = new window.DOMParser();
   const paragraphHtml = parser.parseFromString(attrs.rawHtml || attrs.displayLabel, 'text/html');
+  const marksFromAttrs = translateFieldAttrsToMarks(attrs);
+  const allMarks = [...marks, ...marksFromAttrs]
 
-  const state = EditorState.create({
-    doc: PMDOMParser.fromSchema(params.editorSchema).parse(paragraphHtml),
+  let state = EditorState.create({
+    doc: PMDOMParser.fromSchema(editorSchema).parse(paragraphHtml),
   });
 
+  if (allMarks.length) {
+    state = applyMarksToHtmlAnnotation(state, allMarks);
+  }
+
   const htmlAnnotationNode = state.doc.toJSON();
+  
   return {
     name: 'htmlAnnotation',
     elements: translateChildNodes({
@@ -1822,6 +1830,8 @@ const translateFieldAttrsToMarks = (attrs = {}) => {
     bold,
     underline,
     italic,
+    textColor,
+    textHighlight,
   } = attrs;
 
   const marks = [];
@@ -1830,6 +1840,8 @@ const translateFieldAttrsToMarks = (attrs = {}) => {
   if (bold) marks.push({ type: 'bold', attrs: {} });
   if (underline) marks.push({ type: 'underline', attrs: {} });
   if (italic) marks.push({ type: 'italic', attrs: {} });
+  if (textColor) marks.push({ type: 'color', attrs: { color: textColor } });
+  if (textHighlight) marks.push({ type: 'highlight', attrs: { color: textHighlight } });
   return marks;
 };
 
@@ -1894,6 +1906,34 @@ function translateFieldAnnotation(params) {
             attributes: {
               'xmlns:w': customXmlns,
               'w:val': attrs.multipleImage,
+            },
+          },
+          {
+            name: 'w:fieldFontFamily',
+            attributes: {
+              'xmlns:w': customXmlns,
+              'w:val': attrs.fontFamily,
+            },
+          },
+          {
+            name: 'w:fieldFontSize',
+            attributes: {
+              'xmlns:w': customXmlns,
+              'w:val': attrs.fontSize,
+            },
+          },
+          {
+            name: 'w:fieldTextColor',
+            attributes: {
+              'xmlns:w': customXmlns,
+              'w:val': attrs.textColor,
+            },
+          },
+          {
+            name: 'w:fieldTextHighlight',
+            attributes: {
+              'xmlns:w': customXmlns,
+              'w:val': attrs.textHighlight,
             },
           },
         ],
@@ -2103,3 +2143,55 @@ function resizeKeepAspectRatio(width, height, maxWidth) {
   }
   return { width, height };
 }
+
+function applyMarksToHtmlAnnotation(state, marks) {
+  const { tr, doc, schema } = state;
+  const allowedMarks = ['fontFamily', 'fontSize', 'highlight'];
+
+  if (
+    !marks.some((m) => allowedMarks.includes(m.type))
+  ) {
+    return state;
+  }
+
+  const fontFamily = marks.find((m) => m.type === 'fontFamily');
+  const fontSize = marks.find((m) => m.type === 'fontSize');
+  const highlight = marks.find((m) => m.type === 'highlight');
+
+  const textStyleType = schema.marks.textStyle;
+  const highlightType = schema.marks.highlight;
+
+  doc.descendants((node, pos) => {
+    if (!node.isText) return;
+
+    const foundTextStyle = node.marks.find((m) => m.type.name === 'textStyle');
+    const foundHighlight = node.marks.find((m) => m.type.name === 'highlight');
+    
+    // text style (fontFamily, fontSize)
+    if (!foundTextStyle) {
+      tr.addMark(pos, pos + node.nodeSize, textStyleType.create({
+        ...fontFamily?.attrs,
+        ...fontSize?.attrs,
+      }));
+    } else if (!foundTextStyle?.attrs.fontFamily && fontFamily) {
+      tr.addMark(pos, pos + node.nodeSize, textStyleType.create({
+        ...foundTextStyle?.attrs,
+        ...fontFamily.attrs,
+      }));
+    } else if (!foundTextStyle?.attrs.fontSize && fontSize) {
+      tr.addMark(pos, pos + node.nodeSize, textStyleType.create({
+        ...foundTextStyle?.attrs,
+        ...fontSize.attrs,
+      }));
+    }
+
+    // highlight
+    if (!foundHighlight) {
+      tr.addMark(pos, pos + node.nodeSize, highlightType.create({
+        ...highlight?.attrs,
+      }));
+    }
+  });
+
+  return state.apply(tr);
+};