+++ packages/super-editor/src/core/super-converter/v2/importer/annotationImporter.js
@@ -13,7 +13,8 @@ export const handleAnnotationNode = (params) => {
   const node = nodes[0];
   const sdtPr = node.elements.find((el) => el.name === 'w:sdtPr');
   const sdtContent = node.elements.find((el) => el.name === 'w:sdtContent');
-  const { attrs: marksAsAttrs, marks } = parseAnnotationMarks(sdtContent);
+  const type = sdtPr?.elements.find((el) => el.name === 'w:fieldTypeShort')?.attributes['w:val'];
+  const { attrs: marksAsAttrs, marks } = parseAnnotationMarks(sdtContent, type);
 
   const docPartObj = sdtPr?.elements.find((el) => el.name === 'w:docPartObj');
   if (docPartObj) {
@@ -23,9 +24,12 @@ export const handleAnnotationNode = (params) => {
   const alias = sdtPr?.elements.find((el) => el.name === 'w:alias');
   const tag = sdtPr?.elements.find((el) => el.name === 'w:tag');
   const fieldType = sdtPr?.elements.find((el) => el.name === 'w:fieldType')?.attributes['w:val'];
-  const type = sdtPr?.elements.find((el) => el.name === 'w:fieldTypeShort')?.attributes['w:val'];
   const fieldColor = sdtPr?.elements.find((el) => el.name === 'w:fieldColor')?.attributes['w:val'];
   const isMultipleImage = sdtPr?.elements.find((el) => el.name === 'w:fieldMultipleImage')?.attributes['w:val'];
+  const fontFamily = sdtPr?.elements.find((el) => el.name === 'w:fieldFontFamily')?.attributes['w:val'];
+  const fontSize = sdtPr?.elements.find((el) => el.name === 'w:fieldFontSize')?.attributes['w:val'];
+  const textColor = sdtPr?.elements.find((el) => el.name === 'w:fieldTextColor')?.attributes['w:val'];
+  const textHighlight = sdtPr?.elements.find((el) => el.name === 'w:fieldTextHighlight')?.attributes['w:val'];
 
   const attrs = {
     type,
@@ -34,7 +38,13 @@ export const handleAnnotationNode = (params) => {
     fieldType,
     fieldColor,
     multipleImage: isMultipleImage === 'true',
+    fontFamily: fontFamily !== 'null' ? fontFamily : null,
+    fontSize: fontSize !== 'null' ? fontSize : null,
+    textColor: textColor !== 'null' ? textColor : null,
+    textHighlight: textHighlight !== 'null' ? textHighlight : null,
   };
+
+  const allAttrs = { ...attrs, ...marksAsAttrs };
 
   if (!attrs.fieldId || !attrs.displayLabel) {
     return { nodes: [], consumed: 0 };
@@ -43,14 +53,14 @@ export const handleAnnotationNode = (params) => {
   let result = {
     type: 'text',
     text: `{{${attrs.displayLabel}}}`,
-    attrs: { ...attrs, ...marksAsAttrs },
+    attrs: allAttrs,
     marks,
   };
   
   if (params.editor.options.annotations) {
     result = {
       type: 'fieldAnnotation',
-      attrs: { ...attrs, ...marksAsAttrs }
+      attrs: allAttrs,
     };
   };
   
@@ -65,8 +75,20 @@ export const handleAnnotationNode = (params) => {
  * @param {Object} content The sdtContent node
  * @returns {Object} The attributes object
  */
-export const parseAnnotationMarks = (content = {}) => {
-  const run = content.elements?.find((el) => el.name === 'w:r');
+export const parseAnnotationMarks = (content = {}, type) => {
+  let mainContent = content;
+
+  if (type === 'html') {
+    /// Note: html annotation has a different structure and can include 
+    /// several paragraphs with different styles. We could find the first paragraph 
+    /// and take the marks from there, but we take fontFamily and fontSize from the annotation attributes.
+
+    /// Example: 
+    /// const firstPar = content.elements?.find((el) => el.name === 'w:p');
+    /// if (firstPar) mainContent = firstPar;
+  }
+
+  const run = mainContent.elements?.find((el) => el.name === 'w:r');
   const rPr = run?.elements?.find((el) => el.name === 'w:rPr');
   if (!rPr) return {};