+++ packages/super-editor/src/dev/components/DeveloperPlayground.vue
@@ -45,6 +45,14 @@ const onCreate = ({ editor }) => {
 
   // Set debugging pagination value from editor plugin state
   isDebuggingPagination.value = PaginationPluginKey.getState(editor.state)?.isDebugging;
+
+  // editor.commands.addFieldAnnotation(0, {
+  //   type: 'text',
+  //   displayLabel: 'Some text',
+  //   fieldId: '123',
+  //   fieldType: 'TEXTINPUT',
+  //   fieldColor: '#980043',
+  // });
 };
 
 const onCommentClicked = ({ conversation }) => {
@@ -68,6 +76,7 @@ const editorOptions = computed(() => {
     users: [], // For comment @-mentions, only users that have access to the document
     pagination: true,
     telemetry: telemetry.value,
+    annotations: true,
   }
 });