+++ packages/super-editor/src/extensions/field-annotation/field-annotation-marks-impl.js
@@ -193,6 +193,11 @@ export const FieldAnnotation = Node.create({
       extras: {
         default: {},
         rendered: false,
+      },
+
+      generatorIndex: {
+        default: null,
+        rendered: false,
       }
     };
   },